#38 Use a general bottalking channel ID
Instead of specifying the channel ID for the DingDongListener separately, a general bottalking channel ID can be configured.

+++ app/src/main/kotlin/de/dagadeta/schlauerbot/dingdong/DingDongListener.kt
@@ -37,7 +37,7 @@ class DingDongListener(
     }
 
     override fun onSlashCommandInteraction(event: SlashCommandInteractionEvent) {
-        if (event.channel.id != dingDongConfig.channelId) return
+        if (event.channel.id.toLong() != dingDongConfig.channelId) return
         if (event.name != dingCommand) return
 
         logger.info { "received !ding" }
